Move shared entry point functions to KSPCommonMain
